modify files

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,6 +1,6 @@
 #include "search_algos.h"
 /**
- * linear_search - Searches for a value in an array of integers
+ * linear_search - Searches for a value in an array of integers using the Linear search algorithm
  *
  * @array: The array passed to search in the value in side it
  * @size: The size of the array
@@ -9,12 +9,12 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-    size_t i;
+    int i;
 
     if (array == NULL)
         return (-1);
 
-    for (i = 0; i < size; i++)
+    for (i = 0; i < (int)size; i++)
     {
         printf("Value checked array[%d] = [%d]\n", i, array[i]);
         if (array[i] == value)
